azure-functions/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.13.0
motor>=3.3.1
pymongo>=4.5.0
python-multipart>=0.0.9
fastapi>=0.110.1
python-dotenv>=1.0.1
pydantic>=2.6.4
uvicorn>=0.25.0
aiohttp>=3.9.0
cryptography>=42.0.8
pyjwt>=2.10.1
passlib>=1.7.4
pandas>=2.2.0
numpy>=1.26.0
# AI-specific dependencies
tiktoken>=0.7.0
transformers>=4.36.0
sentence-transformers>=2.5.0
langchain>=0.1.0
spacy>=3.7.0
pypdf>=3.17.0
docx2txt>=0.8
markdown>=3.5.1